+++ AppBox3/A3Categories/NSString+conversion.m
@@ -97,7 +97,7 @@
 }
 
 + (NSString *)orderStringWithOrder:(NSInteger)order {
-	NSString *string = [NSString stringWithFormat:@"%d", order];
+	NSString *string = [NSString stringWithFormat:@"%ld", (long)order];
 	return [string stringByPaddingZeroLength10];
 }
