+++ AppBox3/Common/A3Utilities.h
@@ -12,7 +12,7 @@ void addLeftGradientLayer8Point(UIView *targetView);
 void addRightGradientLayer8Point(UIView *targetView);
 void drawLinearGradient(CGContextRef context, CGRect rect, NSArray *colors);
 
-extern float dash_line_pattern[];
+extern CGFloat dash_line_pattern[];
 
 @interface A3Utilities : NSObject
