+++ AppBox3/Controls/A3DateKeyboardViewController_iPhone.m
@@ -74,7 +74,7 @@
 		[button setTitle:nil forState:UIControlStateNormal];
 		button.mainTitle.text = [monthSymbols objectAtIndex:index];
 		index++;
-		button.subTitle.text = [NSString stringWithFormat:@"%d", index];
+		button.subTitle.text = [NSString stringWithFormat:@"%lu", (unsigned long)index];
 	}
 }
